Details
Fixed details page.  For the select list items, I was sending the Name
field as selected when I should have been sending the Id.  This resulted
in behavior such that when the priority was a a number only, it would be
the selected item, but if it was 1 or 5 which contains "highest" or
"lowest," the system wouldn't recognize it and the topmost list item
would be selected.